Split up classes and added some documentation
