+++ src/Crux.Domain.Persistence/Crux.Domain.Persistence.csproj
@@ -57,6 +57,7 @@
     <Compile Include="ConnectionProviders\SqlConnectionProvider.cs" />
     <Compile Include="NHibernate\Config\SqlExceptionConverter.cs" />
     <Compile Include="NHibernate\Config\SqlExceptionMapper.cs" />
+    <Compile Include="NHibernate\Config\SqlExceptionMapping.cs" />
     <Compile Include="NHibernate\Config\TreatStringAsSqlAnsiString.cs" />
     <Compile Include="ConnectionProviders\IDbConnectionProvider.cs" />
     <Compile Include="NHibernate\NHibernateRepositoryOfId.cs" />
