+++ src/Crux.Domain.Persistence/NHibernate/Config/SqlExceptionMapper.cs
@@ -1,16 +1,23 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
 using System.Linq;
 using NHibernate;
 using NHibernate.Exceptions;
 
 namespace Crux.Domain.Persistence.NHibernate.Config
 {
+    /// <summary>
+    /// This class allows you to customize how <see cref="T:System.Data.Common.DbException"/>s 
+    /// thrown by the underlying data store are converted into NHibernate's 
+    /// exception hierarchy.
+    /// </summary>
     public class SqlExceptionMapper
     {
         private readonly IList<SqlExceptionMapping> _mappings;
 
+        /// <summary>
+        /// Creates a SqlExceptionMapper with default converters for SQL Server.
+        /// </summary>
         public static SqlExceptionMapper Default()
         {
             var mappings = new List<SqlExceptionMapping> {
@@ -31,11 +38,27 @@ namespace Crux.Domain.Persistence.NHibernate.Config
             return new SqlExceptionMapper(mappings);
         }
 
+        /// <summary>
+        /// Creates an empty SqlExceptionMapper with no mappings.
+        /// </summary>
+        public SqlExceptionMapper() : this(new List<SqlExceptionMapping>()) { }
+
+        /// <summary>
+        /// Creates a SqlExceptionMapper with the provided mappings.
+        /// </summary>
+        /// <param name="mappings">An <see cref="IList{T}"/> of mappings 
+        ///     containing the conversion rules.</param>
         public SqlExceptionMapper(IList<SqlExceptionMapping> mappings)
         {
             _mappings = mappings;
         }
 
+        /// <summary>
+        /// Iterates through the provided mappings looking for a matching rule. 
+        /// If found, the converter function for that mapping is called.
+        /// </summary>
+        /// <param name="exceptionContext">Available info about the exception 
+        ///     being thrown.</param>
         public Exception Convert(AdoExceptionContextInfo exceptionContext)
         {
             var mapping = _mappings.FirstOrDefault(m => m.Matcher.Invoke(exceptionContext));
@@ -45,39 +68,28 @@ namespace Crux.Domain.Persistence.NHibernate.Config
                 : SQLStateConverter.HandledNonSpecificException(exceptionContext.SqlException, exceptionContext.Message, exceptionContext.Sql);
         }
 
+        /// <summary>
+        /// Adds a mapping rule to the top of the list.
+        /// </summary>
         public void AddExceptionMappingToTop(SqlExceptionMapping mapping)
         {
             _mappings.Insert(0, mapping);
         }
 
+        /// <summary>
+        /// Adds a mapping rule to the end of the list.
+        /// </summary>
         public void AddExceptionMappingToEnd(SqlExceptionMapping mapping)
         {
             _mappings.Add(mapping);
         }
 
+        /// <summary>
+        /// Clears the internal mappings.
+        /// </summary>
         public void ClearMappings()
         {
             _mappings.Clear();
         }
     }
-
-    public class SqlExceptionMapping
-    {
-        public Func<AdoExceptionContextInfo, bool> Matcher { get; private set; }
-        public Func<AdoExceptionContextInfo, Exception> Converter { get; private set; }
-
-        private SqlExceptionMapping() { }
-
-        public static SqlExceptionMapping MatchOnSqlErrorCode(IEnumerable<int> errorCodes, Func<AdoExceptionContextInfo, Exception> converter)
-        {
-            return new SqlExceptionMapping {
-                Matcher = c => {
-                    var sqlException = ADOExceptionHelper.ExtractDbException(c.SqlException) as SqlException;
-                    return sqlException != null && errorCodes.Contains(sqlException.ErrorCode);
-                },
-                Converter = converter
-            };
-        }
-    }
-
 }
